Adding initial version of mnist tensorflow (note:  original isn't Convolutional Net yet)
This initial version isn't Convolutional on purpose.  The purpose of adding this is to have
a starting version against which we can see what transitioning to a convolutional neural net
looks like.

Fwiw, this initial ML algorithm is using gradient descent to optimize cross entropy.
It's accuracy was in the upper 80s/lower 90s.