Temporary fix for std::bad_alloc error that occurs when evaluating mnist's accuracy on test set.
Fix:  Limit size of test set to 2k (instead of whole 10k).

Longer Explanation:
This python program accumulates a lot of memory as it runs.  Its final
step is to check accuracy of model against mnist.test.images.  This test
set is 10,000 elements large.  When it executes we get an std::bad_alloc
error.

The hypothesis under which this fix is operating is that we're running
out of memory.  It's unclear why we're running out of memory, because
while it's getting into the multiple GB range that's still not so big
that it'd exhaust Windows commit limits.  Could it be that python is
capping it or tensorflow or maybe tensorflow is asking for a contiguous
chunk that's just not available in virtual memory?

Right now I don't know the answer to that, but that's ok for right now
since my primary goal is to better understand neural networks and right
now this memory issue isn't getting into my way.  Hence, we'll just
band-aid fix things for now.