Transformed convolutional ReLU network into basic sigmoid (30 hidden neurons)
Why are we doing this?  The purpose is to get back to a simpler neural
network so that we can better grasp what's going on.  From there we can
do things like add L2 regularization or better weight initialization to
see how that affects things.

1. Deleted all code related to convolutional network unless helpful for
basic sigmoid.

2. Add name param to weight_variable and bias_variable creation so
tensorboard viz can be better annotated.

3. Setup a single hidden layer sigmoid network with 30 hidden neurons.

4. citemsInSet:  Wanted to make mini-batch set size the same.  When one
is much larger than the other, the smaller set lacks granularity into
probability which can make comparing train v. test more difficult (e.g.
test set size of 10 and train set size of 100, we can have 10+
probability differences pretty easily as it only takes one wrong
labeling in set size of 10 to drop score to 90).  This consistency of
set size isn't required if both sets are large-ish (say, 1000+), but for
now while we're doing quick iterations for the sake of tuning over
smaller batches, we'll keep this var in there.
	Init'ing it to 100 because that didn't slow things down much.

5. Reduced number of epochs to 100 because we were getting good progress
in that span.  We needn't waste time with additional calcs.