Adding L2 Regularization of weights to loss function
Currently, I'm not seeing this improve the results much.  In fact, it
might be worsening them.  Easy to turn it off and nice to have the
option, so still worthwhile checking in.

Also updated print statement at the end so that it uses new "format"
based string interpolation, and (more importantly) it will output more
decimals of accuracy calc.